Create Linkedlistframe
The code defines a class named LinkedList which represents a linked list data structure.

The class has a member variable named head of type Node which represents the first node or the head of the linked list.

The class also has an inner class named Node which represents a node in the linked list. The Node class has two member variables: data which stores the data of the node, and next which stores a reference to the next node in the linked list.

The Node class has a constructor which takes an integer value d as an argument and initializes the data member variable with it. The next member variable is by default initialized as null.

Note that since the Node class is an inner class of the LinkedList class, it has access to the head member variable of the LinkedList class. This allows the Node class to easily reference the first node of the linked list.